Fix: Uniform buffers uses size_t as size and deafult frame buffer depth attachment value is 1

+++ KuchCraft/src/Graphics/Core/UniformBuffer.cpp
@@ -29,7 +29,7 @@ namespace KuchCraft {
 		return Ref<UniformBuffer>(new UniformBuffer(size));
 	}
 
-	void UniformBuffer::SetData(const void* data, uint32_t size, uint32_t offset)
+	void UniformBuffer::SetData(const void* data, size_t size, uint32_t offset)
 	{
 		glNamedBufferSubData(m_RendererID, offset, size, data);
 	}
